backend/requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
SQLAlchemy>=2.0
alembic
psycopg[binary]
passlib[argon2]
python-jose[cryptography]
email-validator
redis>=5
anthropic>=0.36.0
openai>=1.0.0
stripe>=7.0.0
python-multipart
psycopg2-binary
boto3>=1.28.0
httpx>=0.24.0
